requirements.txt
streamlit
aiohttp
pandas
pillow
